Make exported API more flexible
The current package API is both rigid and difficult to test; in general,
we should avoid exposing complex functionality only via import-time
side-effects.

This PR refactors the package API to be more testable and flexible, and
then adds a code sample to the README, some testable examples, and
enough tests to get 100% coverage on the main package.