Refactor maxprocs: better logging and only set on success
Currently, we call `GOMAXPROCS` even if we are unable to determine the
CPU quota. Since the returned value is 0, this doesn't have an impact,
but avoid calling `runtime.GOMAXPROCS` to set the value unless we
successfully determine the quota.

Improve logging so users have clear messaging about the source of logs
(maxprocs) and what is being changed (are we ignoring, updating, etc)
and make all the messages more consistent.