Minimum GOMAXPROCS should be 1
We previously set the minimum GOMAXPROCS value to 2 as that was the
minimum allocation we expected internally. However, if the allocation is
actually smaller, we should use lower GOMAXPROCS.

This will also allow us to reduce the allocation size and use a more
appropriate GOMAXPROCS which will reduce the amount of CFS throttling.